+++ app/service/web/pvuvip_task.js
@@ -82,6 +82,9 @@ class PvuvipTaskService extends Service {
                 }, 'pvuvip');
             }
 
+            // 流量峰值 超过历史top邮件触达
+            this.ctx.service.emails.highestPvTipsEmail({ appId, pv, uv, ip, ajax });
+
         } catch (err) { console.log(err); }
     }
